Fix chunk generation, modify hybrid scenario, further improvements
Fixed the logic of util.GenerateChunkFromSet to
create numThreads chunks of size =~ len(set)/numThreads
instead of size =~ numThreads. This spawns significantly
less routines and reduces routine overhead and memory usage.

Minor changes to item-item, tag and title recommenders

Modified findSimilarMovies to enable usage in hybrid model
The current hybrid scenario takes a movieID as input and
computes all similar movies of the requested movie using
item-item, tag-based and title-based collaborative filteting
The results are combined using a sort-merge-join approach,
so only the common similar movies of all 3 methods are considered
in the final rankings.
Adapted UI (index.js) since we dont need to include dataType
in the response of the backend anymore.

Further optimised the configuration mechanism to give more useful
feedback to the user when one of the binaries is not ran properly.